+++ test/components/available-data-list/available-data-list.component.spec.js
@@ -33,4 +33,35 @@ describe('Available data list component', function describe() {
     expect(spy.called).to.be.true;
     wrapper.unmount();
   });
+
+  it('should filter items correctly', function it() {
+    const props = {
+      onSelectItem: () => {},
+      onUnselectItem: () => {},
+      items: List([
+        {
+          uuid: 1,
+          value: 1,
+          label: 'one',
+        },
+        {
+          uuid: 2,
+          value: 2,
+          label: 'two',
+        },
+        {
+          uuid: 3,
+          value: 3,
+          label: 'three',
+        },
+      ]),
+    };
+    const wrapper = mount(<AvailableDataList {...props} />);
+    expect(wrapper.find('.oc-select-order-list-available-data-list').exists()).to.equal(true);
+    expect(wrapper.find('.oc-select-order-list-available-data-item').length).to.equal(3);
+    wrapper.setProps({ searchKeyword: 'two' });
+    wrapper.update();
+    expect(wrapper.find('.oc-select-order-list-available-data-item').length).to.equal(1);
+    wrapper.unmount();
+  });
 });